Use new ref counted memory for pubsub messages too (instead of own ps_priv_t reference counter). Moreover, fixed pubsub messages subscriptions reference: avoid overriding each time msg->sub with new module's subscription when sending messages to subscribed modules; instead, push subscription sources on a queue and de-reference them all-together when message gets destroyed.

+++ Lib/priv.c
@@ -1,7 +1,7 @@
 #include "priv.h"
 
 typedef struct {
-    uint32_t refs;
+    size_t refs;
     ref_dtor dtor;
 } m_header_t;
